pyproject.toml
[tool.poetry]
name = "testovoe_bauart"
version = "0.1.0"
description = ""
authors = ["Гурин Никита Сергеевич <nsgurin@fintechiq.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
clickhouse-driver = "^0.2.4"
minio = "^7.1.10"
aioch = "^0.0.2"
python-dotenv = "^0.20.0"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
python-multipart = "^0.0.5"
pandas = "^1.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
